End-to-End Machine Learning Project in Python
Final project for the Masterclass in Business Analytics module. 

This work aims to build Machine Learning models that can be used by used-cars dealership owners to help them make data-driven decisions and improve their pricing strategy. Two models are presented, one that predicts the true market price of a used car and another model that predicts the time it takes to sell a specific car given its attributes. The chosen dataset for this coursework was the eBay Used Cars Database. This dataset was downloaded from the platform Kaggle and consists of over 370,000 observations on used car sales from eBay Germany. We used Python and Jupyter Notebook to develop our algorithms. Our group found that Random Forest and Neural Network are the recommended models to predict market price, and Neural Network is the recommended model to predict the number of days it takes to sell a specific car.